moved class structure
removed class dependencies, made functions work without a self instance, fixed class structure
