+++ src/sm/synthesiser/synthesiser.py
@@ -1,35 +1,946 @@
-from pydantic import BaseModel
-from typing import Dict, Any, List
-import random
-from sm.synthesiser.initialiser.get_constraints import get_constraints_class
-from sm.synthesiser.initialiser.match_methods import match_methods_class
-from sm.synthesiser.field_match.match import match_class
-from sm.synthesiser.generator.constraint_based_generator import constraint_based_generator_class
-from sm.synthesiser.generator.synth_based_generator import synth_based_generator_class
+from typing import (
+    List,
+    Dict,
+    Tuple,
+    Set,
+    Union,
+    Optional,
+    Literal,
+    Any,
+    Annotated,
+    Type,
+    get_args,
+    get_origin,
+)
+from sm.synthesiser.helper_funcs.constraints import make_one_string, make_one_decimal, make_new_contraints, recursive_get_applied_constraints
+from sm.synthesiser.helper_funcs.matching_fields import match_fields, recursive_match_fields
+from sm.synthesiser.helper_funcs.provider_methods import list_match_methods, make_resolved_methods
 from sm.tools.model_funcs import get_model_fields
 from sm.pre_made_data import provider_methods
+from sm.tools.regex_generator import regex_builder
+from collections import deque
+from multiprocessing import Pool
+from functools import partial
+from pydantic import BaseModel
+from decimal import Decimal, ROUND_HALF_UP
+from sm.pre_made_data import (
+    all_constr_attribs,
+    default_constr_dict,
+    recursive_types,
+    python_builtin_types,
+)
+import re
+import time
+import string
+import random
+import rstr
+import inspect
+import exrex
+
+
 
 """
-The main synthesiser class
-pulls from sub classes to build full functionality
+Generation path:\n
+    format of  'Schema_name(field_name)[amount/length]'
+    format of  'n(f)[a].Type(index)[amount/length]'
+    e.g.
+    1:\n
+        class User():
+            address: (_,_,_)
+        User(address)[1].Tuple(0)[1]
+        -> would be generating a data pool for
+        class User():
+            address: (X,_,_)
+    2:\n
+        class Address():
+            street : str
+        class User():
+            nested : Address
+        User(nested)[100].Address(street)[1]
+    3:\n
+        class User():
+            address: List[str,int]
+        User(address)[1].List(0)[10]
+        -> means
+            [str]*10
+        
+        User(address)[1].List(1)[20]
+        -> means
+            [int]*20
 """
 
 
-class Synthesiser(
-    get_constraints_class,
-    match_methods_class,
-    match_class,
-    constraint_based_generator_class,
-    synth_based_generator_class,
-):
+def print_path(generate_path, elapsed_time):
+    """
+    inputs:
+        generate path
+        elapsed time
+    prints based on path depth:
+        Time taken: ... seconds          path
+    """
+    pooling_depth = len(list(map(int, re.findall(r"\[(\d+)\]", generate_path))))
+    time_str = f"{elapsed_time:.2f}"
+    print(
+        f"Time taken: {time_str} seconds",
+        " " * (10 - max(10, len(time_str))),
+        "    " * (pooling_depth - 1),
+        generate_path,
+    )
+
+
+string_list = string.ascii_letters + string.digits
+
+
+
+class Synthesiser():
     def __init__(self, method="faker"):
         self.outputpooling = {}
         data_set = provider_methods[method]
-        self.word_list, methods_map = self.list_match_methods(method)
+        self.word_list, methods_map = list_match_methods(method)
         self.word_tokens = data_set['word_tokens'] # {word: word.split("_") for word in word_list}
         self.word_tokens_set = data_set['word_tokens_set'] # {word: set(word.split("_")) for word in word_list}
-        self.resolved_methods = self.make_resolved_methods(self.word_list, methods_map)
-    
+        self.resolved_methods = make_resolved_methods(self.word_list, methods_map)
+
+    # ----- Constraint based generation -----
+    def generate_from_constraints(
+        self, field_name:str, constraints:Dict[str,Any], generate_path:str
+    ) -> Any:  # value
+        """
+        Inputs:\n
+            field_name
+            constraints ( generated by check_generation_constraints() )
+            generation path string
+        Outputs:\n
+            a single generated value based on constraints
+            if a data pool for the generation path hasnt been made:
+                it will first generate a data pool for the path
+        """
+
+        # print(f"Name:{field_name}\nFields:{constraints}")
+        # print("__")
+
+        if generate_path not in self.outputpooling or (
+            generate_path in self.outputpooling
+            and self.outputpooling[generate_path] == []
+        ):
+            pooling_numbers = list(map(int, re.findall(r"\[(\d+)\]", generate_path)))
+            amount = pooling_numbers[0]
+            pooling_count = 1
+            for x in pooling_numbers:
+                pooling_count *= x
+
+            # print(f"Path:{generate_path}")
+            # print("count", pooling_count)
+            start_time = time.time()
+            data_pool = []
+
+            if constraints["pattern"] is not None:
+                # print(f"Path:{generate_path}")
+                # HERE HERE
+                # check if a map exists for this constraint sequence (pattern and constraints)
+                # if there is not, generate a map of possible results and then run regex forward pass over each (could be slow for initial pass through)
+                # save the forward pass map to global
+
+                # print(constraints["pattern"])
+
+                min_length = constraints["min_length"]
+                max_length = constraints["max_length"]
+                pattern = constraints["pattern"]
+                if min_length is None and max_length is None:
+                    pattern = pattern
+                elif min_length is not None and max_length is None:
+                    # pattern = f"^(?=.{{{min_length},}}$)(?:{pattern})$"
+                    if pattern[-1] == "$":
+                        pattern = pattern[:-1]
+                    while pattern[-1] in ["+", "*", "?"]:
+                        pattern = pattern[:-1]
+                    pattern = pattern + f"{{{min_length},}}$"
+
+                elif min_length is None and max_length is not None:
+                    # pattern = f"^(?=.{{0,{max_length}}}$)(?:{pattern})$"
+                    if pattern[-1] == "$":
+                        pattern = pattern[:-1]
+                    while pattern[-1] in ["+", "*", "?"]:
+                        pattern = pattern[:-1]
+                    pattern = pattern + f"{{0,{max_length}}}$"
+                else:
+                    # pattern = f"^(?=.{{{min_length},{max_length}}}$)(?:{pattern})$"
+                    if pattern[-1] == "$":
+                        pattern = pattern[:-1]
+                    while pattern[-1] in ["+", "*", "?"]:
+                        pattern = pattern[:-1]
+                    pattern = pattern + f"{{{min_length},{max_length}}}$"
+
+                # print(pattern)
+                # print(exrex.getone(pattern))
+                # print(self.make_one_string(pattern))
+                PERFORMANCE = True  # risky but faster
+                try:
+                    if (max(1, pooling_count) > 10000 and PERFORMANCE) or ('(?' in pattern):
+                        src = regex_builder.compile_regex_to_function_source(pattern)
+                        safe_builtins = {
+                            "len": len, "range": range, "min": min, "max": max,
+                            "list": list, "tuple": tuple, "deque": deque, "chr": chr, "ord": ord, "set": set, "map": map,
+                            "int": int, "AssertionError": AssertionError
+                        }
+                        env = {
+                            "__builtins__": safe_builtins,  # dissallow anything except random and string
+                            "random": random,
+                            "string": string,
+                        }
+                        exec(src, env)
+                        gen = env["gen"]
+                        """
+                        with Pool() as p:
+                            data_pool = p.map(
+                                _worker_compile_and_gen, 
+                                [(pattern, i) for i in range(pooling_count)]
+                            )
+                        #"""
+                        data_pool = [gen() for _ in range(max(1, pooling_count))]
+                    elif max(1, pooling_count) > 100:
+                        with Pool() as p:
+                            data_pool = p.map(
+                                make_one_string, [pattern] * max(1, pooling_count)
+                            )
+                    else:
+                        data_pool = [
+                            exrex.getone(pattern) for x in range(max(1, pooling_count))
+                        ]
+                except Exception as e:
+                    print(f"failed reg pool and used fallback: {e}")
+                    data_pool = [
+                        rstr.xeger(pattern) for x in range(max(1, pooling_count))
+                    ]
+
+                if constraints["to_upper"] is not None:
+                    data_pool = [item.upper() for item in data_pool]
+                elif constraints["to_lower"] is not None:
+                    data_pool = [item.lower() for item in data_pool]
+                if data_pool == []:
+                    raise Exception(f"No data pool for {constraints['pattern']}")
+            else:
+                # print("Generate from annotation",str(constraints["annotation"]))
+                data_type = constraints["annotation"]
+                if data_type is str:
+                    temp_string_list = string_list
+                    if constraints["to_upper"] is not None:
+                        temp_string_list = [item.upper() for item in temp_string_list]
+                    elif constraints["to_lower"] is not None:
+                        temp_string_list = [item.lower() for item in temp_string_list]
+                    if constraints["min_length"] is not None:
+                        min_length = constraints["min_length"]
+                    else:
+                        min_length = 1
+                    if constraints["max_length"] is not None:
+                        max_length = constraints["max_length"]
+                    else:
+                        max_length = min_length + max(1, (pooling_count // amount)) + 3
+                    data_pool = [
+                        "".join(
+                            random.choices(
+                                temp_string_list,
+                                k=(random.randint(min_length, max_length)),
+                            )
+                        )
+                        for x in range(max(1, pooling_count))
+                    ]
+
+                    if data_pool == []:
+                        raise Exception(f"No data pool for str:{generate_path}")
+
+                elif data_type is int or data_type is float or data_type is Decimal:
+                    if constraints["lt"] is not None:
+                        lt = constraints["lt"] - 1
+                    else:
+                        lt = 10 * max(1, (pooling_count // amount) + 3)
+                    if constraints["le"] is not None:
+                        lt = min(lt, constraints["le"])
+
+                    if constraints["gt"] is not None:
+                        gt = constraints["gt"] + 1
+                    else:
+                        gt = lt * -1
+                    if constraints["ge"] is not None:
+                        gt = max(gt, constraints["ge"])
+
+                    if gt > lt:
+                        raise Exception(
+                            f"Value for gt:{gt} is greater than value for lt:{lt}"
+                        )
+
+                    if constraints["multiple_of"]:
+                        multiple_of = constraints["multiple_of"]
+                        first = ((gt + multiple_of) // multiple_of) * multiple_of
+
+                        if data_type is int or (
+                            (data_type is float)
+                            and constraints["decimal_places"] is None
+                        ):
+                            count = (lt - gt) // multiple_of
+                            if count <= 0:
+                                raise Exception(
+                                    f"No multiples of {multiple_of} fit in the range [{gt}, {lt})."
+                                )
+                            try:
+                                data_pool = [
+                                    first + (idx + 1) * multiple_of
+                                    for idx in range(count - 1)
+                                ]  # count is capped at poling_count
+                            except:
+                                data_pool == []
+                        else:  # data_type is float and constraints["decimal_places"] is not None
+                            decimal_places = constraints["decimal_places"]
+                            scale = Decimal(10) ** Decimal(decimal_places)
+                            decimal_precision = Decimal(10) ** Decimal(
+                                decimal_places * -1
+                            )
+                            # print(str(multiple_of))
+                            # print(Decimal(repr(multiple_of)).as_tuple())
+                            precision = Decimal(10) ** (
+                                Decimal(repr(multiple_of)).as_tuple().exponent
+                            )
+                            # print(precision)
+
+                            first = (
+                                (Decimal(gt) + Decimal(multiple_of))
+                                // Decimal(multiple_of)
+                            ) * Decimal(multiple_of)
+
+                            first = Decimal(first).quantize(
+                                precision, rounding=ROUND_HALF_UP
+                            )
+                            multiple_of = Decimal(multiple_of).quantize(
+                                precision, rounding=ROUND_HALF_UP
+                            )
+
+                            min_scaled = Decimal(first * scale)
+                            scaled_mult = Decimal(multiple_of * scale)
+
+                            # print(first,multiple_of,min_scaled,scaled_mult,scale)
+
+                            data_pool = []
+                            try:
+                                if max(1, pooling_count) > 100:
+                                    with Pool() as p:
+                                        func = partial(
+                                            make_one_decimal,
+                                            decimal_precision=decimal_precision,
+                                            min_scaled=min_scaled,
+                                            scaled_mult=scaled_mult,
+                                            scale=scale,
+                                        )
+                                        data_pool = p.map(
+                                            func,
+                                            range(
+                                                int(
+                                                    (lt * scale - min_scaled)
+                                                    // scaled_mult
+                                                )
+                                                - 1
+                                            ),
+                                        )
+                                    data_pool = [d for d in data_pool if d is not None]
+                                else:
+                                    for x in range(
+                                        int((lt * scale - min_scaled) // scaled_mult)
+                                        - 1
+                                    ):
+                                        potential_val = (
+                                            min_scaled + (x + 1) * scaled_mult
+                                        ) / scale
+                                        if potential_val == potential_val.quantize(
+                                            decimal_precision, rounding=ROUND_HALF_UP
+                                        ):
+                                            data_pool.append(potential_val)
+                            except Exception:
+                                data_pool = []
+                            if data_pool == []:
+                                raise Exception(
+                                    f"No multiples of {multiple_of} fit in the range [{gt}, {lt}) with max decimal digits{decimal_places}."
+                                )
+                    else:
+                        if constraints["decimal_places"] is not None:
+                            data_pool = [
+                                round(
+                                    random.uniform(gt, lt),
+                                    constraints["decimal_places"],
+                                )
+                                for _ in range(max(1, pooling_count))
+                            ]
+                        else:
+                            data_pool = [
+                                random.randint(gt, lt)
+                                for _ in range(max(1, pooling_count))
+                            ]
+
+                    if constraints["allow_inf_nan"] is not None:
+                        allowed_inf_types = ["inf", "-inf", "nan"]
+                        if (
+                            constraints["gt"] is not None
+                            or constraints["ge"] is not None
+                        ):
+                            allowed_inf_types.remove("-inf")
+                            allowed_inf_types.remove("nan")
+                        if (
+                            constraints["lt"] is not None
+                            or constraints["le"] is not None
+                        ):
+                            allowed_inf_types.remove("inf")
+                            if "nan" in allowed_inf_types:
+                                allowed_inf_types.remove("nan")
+                        if constraints["multiple_of"]:
+                            allowed_inf_types = []
+                        if constraints["max_digits"]:
+                            allowed_inf_types = []
+
+                        if data_type is float and allowed_inf_types != []:
+                            data_pool.extend(
+                                [
+                                    random.choice(allowed_inf_types)
+                                    for x in range(len(data_pool))
+                                ]
+                            )
+                    random.shuffle(data_pool)
+                    if data_pool == []:
+                        raise Exception(f"No data pool for num:{generate_path}")
+
+                    # if the data pool doesnt have enough elements, extend it with copies of existing elements
+                    original_len = len(data_pool)
+                    while len(data_pool) < pooling_count:
+                        # Append elements from the original list in order
+                        for item in data_pool[:original_len]:
+                            if len(data_pool) < pooling_count:
+                                data_pool.append(item)
+                            else:
+                                break
+
+                elif data_type is bool:
+                    data_pool = [
+                        random.randint(1, 2) == 1 for x in range(max(1, pooling_count))
+                    ]
+                    if data_pool == []:
+                        raise Exception(f"No data pool for bool:{generate_path}")
+                elif data_type is complex:
+                    # return_value= '0+0j'
+                    max_num = 10 ** max(1, (pooling_count // amount) + 3)
+                    min_num = max_num * -1
+
+                    data_pool = [
+                        complex(
+                            random.randint(0, min_num + max_num) - min_num,
+                            random.randint(0, min_num + max_num) - min_num,
+                        )
+                        for x in range(max(1, pooling_count))
+                    ]
+                    if data_pool == []:
+                        raise Exception(f"No data pool for complex:{generate_path}")
+                elif data_type is bytes:
+                    if constraints["min_length"] is not None:
+                        min_length = constraints["min_length"]
+                    else:
+                        min_length = 1
+                    if constraints["max_length"] is not None:
+                        max_length = constraints["max_length"]
+                    else:
+                        max_length = min_length + max(1, (pooling_count // amount)) + 3
+                    data_pool = [
+                        random.randbytes(random.randint(min_length, max_length))
+                        for x in range(max(1, pooling_count))
+                    ]
+                    if data_pool == []:
+                        raise Exception(f"No data pool for bytes:{generate_path}")
+                else:
+                    data_pool = ["error" for x in range(max(1, pooling_count))]
+                    if data_pool == []:
+                        raise Exception(f"No data pool for default:{generate_path}")
+
+            elapsed_time = time.time() - start_time  # end timer
+            print_path(generate_path, elapsed_time)
+
+            self.outputpooling[generate_path] = data_pool
+            # print(f"Pool | Time taken: {elapsed_time:.2f} seconds\n")
+        return_value = self.outputpooling[generate_path].pop()
+        if self.outputpooling[generate_path] == []:
+            del self.outputpooling[generate_path]
+
+        if constraints["annotation"] is bytes:
+            return_value = constraints["annotation"](str(return_value), "utf-8")
+        else:
+            return_value = constraints["annotation"](return_value)
+        # input("wait...")
+        # exit()
+        return return_value
+
+    def apply_constraints(
+        self,
+        return_value:Any,
+        constraints:Dict[str,Any],
+        match_name:str,
+        generate_path:str,
+        pooling_count:int,
+        amount:int,
+    ) -> Any:
+        """
+        Input:\n
+            value
+            value contraints
+            generation provider name
+            generation path
+            path amount
+            batch amount
+        Output:\n
+            value with constraints applied to it
+        """
+        data_type = constraints["annotation"]
+        if data_type is str:
+            # print("\tconstraining str")
+            temp_string_list = string_list
+            if constraints["to_upper"] is not None:
+                temp_string_list = [item.upper() for item in string_list]
+            elif constraints["to_lower"] is not None:
+                temp_string_list = [item.lower() for item in string_list]
+
+            if constraints["to_upper"] is not None:
+                return_value = return_value.upper()
+            elif constraints["to_lower"]:
+                return_value = return_value.lower()
+            if constraints["min_length"] is not None:
+                min_length = constraints["min_length"]
+            else:
+                min_length = 1
+            if constraints["max_length"] is not None:
+                max_length = constraints["max_length"]
+
+            if len(return_value) < min_length:
+                pad_length = min_length - len(return_value)
+                return_value += "".join(random.choices(temp_string_list, k=pad_length))
+            if constraints["max_length"] is not None:
+                if len(return_value) > max_length:
+                    return_value = return_value[:max_length]
+
+        elif data_type is int or data_type is float:
+            if constraints["lt"] is not None:
+                lt = constraints["lt"] - 1
+            else:
+                lt = 10 * max(1, (pooling_count // amount) + 3)
+            if constraints["le"] is not None:
+                lt = min(lt, constraints["le"])
+
+            if constraints["gt"] is not None:
+                if data_type is int:
+                    gt = constraints["gt"] + 1
+                else:
+                    if constraints["multiple_of"] is not None:
+                        gt = constraints["gt"] + constraints["multiple_of"]
+                    else:
+                        gt = constraints["gt"] + 0.0001
+
+            else:
+                gt = lt * -1
+            if constraints["ge"] is not None:
+                gt = max(gt, constraints["ge"])
+
+            if gt > lt:
+                raise Exception(f"Value for gt:{gt} is greater than value for lt:{lt}")
+
+            if constraints["multiple_of"]:
+                multiple_of = constraints["multiple_of"]
+                first = ((gt + multiple_of) // multiple_of) * multiple_of
+
+                if data_type is int or (
+                    (data_type is float) and constraints["decimal_places"] is None
+                ):
+                    count = (lt - gt) // multiple_of
+                    if count <= 0:
+                        raise Exception(
+                            f"No multiples of {multiple_of} fit in the range [{gt}, {lt})."
+                        )
+                    idx = (return_value // multiple_of) % count
+
+                    return_value = first + idx * multiple_of
+                else:  # data_type is float and constraints["decimal_places"] is not None
+                    return_value = self.generate_from_constraints(
+                        match_name, constraints, generate_path
+                    )
+            else:
+                if constraints["decimal_places"] is not None:
+                    return_value = round(return_value, constraints["decimal_places"])
+                else:
+                    try:
+                        return_value = int(return_value)
+                    except:
+                        return_value = self.generate_from_constraints(
+                            match_name, constraints, generate_path
+                        )
+
+        elif data_type is bool:
+            try:
+                return_value = bool(return_value)
+            except:
+                return_value = random.choice([True, False])
+
+        elif data_type is complex:
+            try:
+                return_value = complex(return_value)
+            except:
+                return_value = self.generate_from_constraints(
+                    match_name, constraints, generate_path
+                )
+
+        elif data_type is bytes:
+            try:
+                return_value = bytes(return_value)
+            except:
+                return_value = self.generate_from_constraints(
+                    match_name, constraints, generate_path
+                )
+        if constraints["annotation"] is bytes:
+            return_value = constraints["annotation"](str(return_value), "utf-8")
+        else:
+            return_value = constraints["annotation"](return_value)
+        return return_value
+    # ----- Constraint based generation -----
+
+    # ----- Synth based generation -----  
+    def generate_synth_data(
+        self, field_name:str, match_name:str, applied_constraints:Dict[str,Any], generate_path:str
+    ) -> Any:
+        """
+        Inputs:\n
+            field name
+            match name
+            constraints
+            generate path
+        Recursive function that crawls through the constraints structure\n
+        Handles collection types and model recursion\n
+        Outputs:\n
+            python structure of the constraints format with fully generated data
+        """
+        # print("__")
+        # print(f"	Generating for: {field_name}")
+        # print(applied_constraints)
+        # data_type = get_origin(applied_constraints["annotation"])
+        data_type = applied_constraints["origin"]
+        output_data = ""
+        if data_type == Annotated:
+            # print(applied_constraints)
+            # print("~~")
+            new_info = get_args(applied_constraints["annotation"])
+            new_data_type = new_info[0]  # [0] is the data type
+            # print(new_data_type)
+            constr_constraints = {}
+
+            for attr in all_constr_attribs:
+                return_val = getattr(
+                    new_info[1], attr, None
+                )  # [1] is the data constraints
+                constr_constraints[attr] = return_val
+            # print(constr_constraints)
+            new_constraints = constr_constraints
+            new_constraints["required"] = True
+            new_constraints["annotation"] = new_data_type
+            new_constraints["origin"] = get_origin(new_data_type)
+            new_constraints["args"] = get_args(new_data_type)
+            # print("___")
+            # print(new_constraints)
+            generate_path += ".Annotated"
+            output_data = self.generate_synth_data(
+                field_name, match_name, new_constraints, generate_path
+            )
+            # print("returned:",output_data)
+
+        else:
+            if not data_type:  # if data_type is None:
+                data_type = applied_constraints["annotation"]
+            data_args = applied_constraints["args"]
+            # data_args = get_args(applied_constraints["annotation"])
+            # print(f"	{data_type}\n	{data_args}")
+            if data_type is type(None):
+                output_data = None
+            elif data_type in recursive_types:
+                if applied_constraints["min_length"] is not None:
+                    min_amount = applied_constraints["min_length"]
+                else:
+                    min_amount = 1
+                if applied_constraints["max_length"] is not None:
+                    max_amount = applied_constraints["max_length"]
+                else:
+                    max_amount = min_amount + 4
+
+                new_applied_constraints = make_new_contraints(applied_constraints)
+
+                if data_type in [List, list]:
+                    output_data = []
+                    if len(data_args) == 0:
+                        data_args = [str]
+                    for x in range(random.randint(min_amount, max_amount)):
+                        chosen_index = random.randint(0, len(data_args) - 1)
+                        chosen_type = data_args[chosen_index]
+                        new_applied_constraints["annotation"] = chosen_type
+                        new_applied_constraints["origin"] = get_origin(chosen_type)
+                        new_applied_constraints["args"] = get_args(chosen_type)
+                        list_generate_path = (
+                            generate_path + f".List({chosen_index})[{max_amount}]"
+                        )
+                        output_data.append(
+                            self.generate_synth_data(
+                                field_name,
+                                match_name,
+                                new_applied_constraints,
+                                list_generate_path,
+                            )
+                        )
+
+                elif data_type in [Dict, dict]:
+                    new_left_applied_constraints = new_applied_constraints.copy()
+                    new_right_applied_constraints = new_applied_constraints.copy()
+                    if len(data_args) == 0:
+                        data_args = [str, str]
+                    chosen_type_left = data_args[0]
+                    new_left_applied_constraints["annotation"] = chosen_type_left
+                    new_left_applied_constraints["origin"] = get_origin(
+                        chosen_type_left
+                    )
+                    new_left_applied_constraints["args"] = get_args(chosen_type_left)
+                    chosen_type_right = data_args[1]
+                    new_right_applied_constraints["annotation"] = chosen_type_right
+                    new_right_applied_constraints["origin"] = get_origin(
+                        chosen_type_right
+                    )
+                    new_right_applied_constraints["args"] = get_args(chosen_type_right)
+                    output_data = {}
+                    current_amount = 0
+                    target_amount = random.randint(min_amount, max_amount)
+                    current_tries = 0
+                    target_amountx2 = target_amount * 2
+                    # for x in range(random.randint(min_amount,max_amount)):
+                    dict_keys = set()
+                    while current_amount < target_amount:
+                        generate_path_v1 = generate_path + f".Dict(Left)[{max_amount}]"
+                        v1 = self.generate_synth_data(
+                            field_name,
+                            match_name,
+                            new_left_applied_constraints,
+                            generate_path_v1,
+                        )
+
+                        dict_keys.add(v1)
+                        current_amount = len(dict_keys)
+                        current_tries += 1
+                        if current_tries > target_amountx2:
+                            if min_amount != 1:
+                                raise Exception(
+                                    f"Not enough provider keys for {field_name}, \nkeys: {dict_keys}"
+                                )
+                            else:
+                                break
+                    for key in dict_keys:
+                        generate_path_v2 = generate_path + f".Dict(Right)[{max_amount}]"
+                        v2 = self.generate_synth_data(
+                            field_name,
+                            match_name,
+                            new_right_applied_constraints,
+                            generate_path_v2,
+                        )
+                        output_data[key] = v2
+
+                elif data_type in [Tuple, tuple]:
+                    output_data = []
+                    if len(data_args) == 0:
+                        data_args = [
+                            str for x in range(random.randint(min_amount, max_amount))
+                        ]
+                    for x in range(len(data_args)):
+                        chosen_type = data_args[x]
+                        new_applied_constraints["annotation"] = chosen_type
+                        new_applied_constraints["origin"] = get_origin(chosen_type)
+                        new_applied_constraints["args"] = get_args(chosen_type)
+                        tuple_generate_path = (
+                            generate_path + f".Tuple({x})[{max_amount}]"
+                        )
+                        output_data.append(
+                            self.generate_synth_data(
+                                field_name,
+                                match_name,
+                                new_applied_constraints,
+                                tuple_generate_path,
+                            )
+                        )
+
+                elif data_type in [Set, set, frozenset]:
+                    output_data = []
+                    current_amount = 0
+                    target_amount = random.randint(min_amount, max_amount)
+                    if len(data_args) == 0:
+                        data_args = [str]
+                    chosen_type = data_args[0]
+                    new_applied_constraints["annotation"] = chosen_type
+                    new_applied_constraints["origin"] = get_origin(chosen_type)
+                    new_applied_constraints["args"] = get_args(chosen_type)
+                    current_tries = 0
+                    target_amountx2 = target_amount * 2
+                    if data_type is frozenset:
+                        pathstr = "FrozenSet"
+                    else:
+                        pathstr = "Set"
+                    while current_amount < target_amount:
+                        set_generate_path = (
+                            generate_path + f".{pathstr}({0})[{max_amount}]"
+                        )
+                        output_data.append(
+                            self.generate_synth_data(
+                                field_name,
+                                match_name,
+                                new_applied_constraints,
+                                set_generate_path,
+                            )
+                        )
+                        current_amount = len(output_data)
+                        current_tries += 1
+                        if current_tries > target_amountx2:
+                            if min_amount != 1:
+                                raise Exception(
+                                    f"Not enough provider keys for {field_name}"
+                                )
+                            else:
+                                break
+                    if data_type is frozenset:
+                        output_data = frozenset(output_data)
+
+                elif data_type == Union:
+                    chosen_index = random.randint(0, len(data_args) - 1)
+                    chosen_type = data_args[chosen_index]
+                    new_applied_constraints["annotation"] = chosen_type
+                    new_applied_constraints["origin"] = get_origin(chosen_type)
+                    new_applied_constraints["args"] = get_args(chosen_type)
+                    generate_path += f".Union({chosen_index})"
+                    output_data = self.generate_synth_data(
+                        field_name, match_name, new_applied_constraints, generate_path
+                    )
+
+                elif data_type == Optional:
+                    # should use Union, but still here for fallback
+                    data_args.extend(None)
+                    chosen_index = random.randint(0, len(data_args) - 1)
+                    chosen_type = data_args[chosen_index]
+                    if chosen_type is None:
+                        output_data = None
+                    else:
+                        new_applied_constraints["annotation"] = chosen_type
+                        new_applied_constraints["origin"] = get_origin(chosen_type)
+                        new_applied_constraints["args"] = get_args(chosen_type)
+                        generate_path += f".Optional({chosen_index})"
+                        output_data = self.generate_synth_data(
+                            field_name,
+                            match_name,
+                            new_applied_constraints,
+                            generate_path,
+                        )
+
+                elif data_type == Literal:
+                    output_data = random.choice(data_args)
+
+                else:
+                    raise Exception(
+                        f"Recersive data type| {data_type} : {data_args} |not handled"
+                    )
+            elif data_type in python_builtin_types:
+                if match_name == "":
+                    func = None
+                else:
+                    func = self.resolved_methods[match_name]
+                if applied_constraints["pattern"] or not func or match_name == "":
+                    output_data = self.generate_from_constraints(
+                        field_name, applied_constraints, generate_path
+                    )
+                else:
+                    if generate_path not in self.outputpooling or (
+                        generate_path in self.outputpooling
+                        and self.outputpooling[generate_path] == []
+                    ):
+                        pooling_numbers = list(
+                            map(int, re.findall(r"\[(\d+)\]", generate_path))
+                        )
+                        pooling_count = 1
+                        for x in pooling_numbers:
+                            pooling_count *= x
+
+                        start_time = time.time()
+                        data_temp_pool = [func() for _ in range(pooling_count)]
+
+                        data_pool = [
+                            self.apply_constraints(
+                                func_val,
+                                applied_constraints,
+                                match_name,
+                                generate_path,
+                                pooling_count,
+                                pooling_numbers[0],
+                            )
+                            for func_val in data_temp_pool
+                        ]
+
+                        elapsed_time = time.time() - start_time
+
+                        print_path(generate_path, elapsed_time)
+
+                        self.outputpooling[generate_path] = data_pool
+                    output_data = self.outputpooling[generate_path].pop()
+
+            else:
+                if inspect.isclass(data_type) and (
+                    issubclass(data_type, BaseModel) or isinstance(data_type, BaseModel)
+                ):
+                    output_data = self.synthesise_recursive(
+                        data_type, self.method, amount=1, path=generate_path + "."
+                    )
+                    # print("big nested")
+
+                else:
+                    raise Exception(
+                        f"Unkown data type ({data_type}) for field {field_name}"
+                    )
+        # print(f"Data: {output_data}")
+        # print("__")
+        # apply constraints of output after data is provided
+        return output_data
+
+    def generate_single_value(self, field_name:str, field_type:Type) -> Any:
+        """
+        Inputs:\n
+            field_name
+            field_type
+        Outputs:\n
+            single generated value
+        """
+        matched_field = match_fields([field_name])
+        if matched_field[field_name] != "":
+            func = self.resolved_methods[matched_field[field_name]]
+            value = func()
+            try:
+                value = field_type(value)
+            except:
+                applied_constraints = default_constr_dict.copy()
+                applied_constraints["annotation"] = field_type
+                value = self.apply_constraints(
+                    value,
+                    applied_constraints,
+                    matched_field[field_name],
+                    "self()[1].generate",
+                    1,
+                    1,
+                )
+        else:
+            applied_constraints = default_constr_dict.copy()
+            applied_constraints["annotation"] = field_type
+            value = self.generate_from_constraints(
+                field_name, applied_constraints, "self()[1].generate"
+            )
+        return value
+    # ----- Constraint based generation -----
+
+    # ----- Central called functions -----
     def synthesise_recursive(
         self, schema_model:BaseModel, method="faker", amount:int=1, path:str=""
     ) -> Dict[str, Any]:
@@ -109,8 +1020,8 @@ class Synthesiser(
 
         self.method = method
 
-        self.field_match_pairs = self.recursive_match_fields(schema_model)
-        self.applied_constraints = self.recursive_get_applied_constraints(schema_model)
+        self.field_match_pairs = recursive_match_fields(schema_model,method)
+        self.applied_constraints = recursive_get_applied_constraints(schema_model)
         # print(self.applied_constraints)
         dataset = []
         for x in range(amount):
@@ -126,3 +1037,4 @@ class Synthesiser(
                 )
         print(f"Completed: {amount}/{amount}{' ' * 30}")
         return dataset
+    # ----- Central called functions -----
